doc updates + minor code move
